fix : 리스폰스 값들의 형변환으로 Double에 구현

+++ UpStock/UpStock/Util/Extension/Int + NumberFormatting.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-extension Int {
+extension Double {
     
     var formattedWithSeparator: String {
         let numberFormatter = NumberFormatter()
